Move the background image in terminal view
